Fix edit question options journey (#317)
* Fix edit question options journey

* Split getServerSideProps into separate file

---------
