+++ packages/admin/src/pages/build-application/[applicationId]/[sectionId]/[questionId]/edit/question-options.page.tsx
@@ -1,180 +1,12 @@
-import {
-  Button,
-  FlexibleQuestionPageLayout,
-  TextInput,
-  ValidationError,
-} from 'gap-web-ui';
-import { GetServerSideProps } from 'next';
-import getConfig from 'next/config';
+import { Button, FlexibleQuestionPageLayout, TextInput } from 'gap-web-ui';
 import { toWordsOrdinal } from 'number-to-words';
 import CustomLink from '../../../../../../components/custom-link/CustomLink';
 import Meta from '../../../../../../components/layout/Meta';
-import { getApplicationFormSummary } from '../../../../../../services/ApplicationService';
-import {
-  getQuestion,
-  patchQuestion,
-} from '../../../../../../services/QuestionService';
-import {
-  ApplicationFormQuestion,
-  ApplicationFormSummary,
-} from '../../../../../../types/ApplicationForm';
-import { QuestionWithOptionsSummary } from '../../../../../../types/QuestionSummary';
-import callServiceMethod from '../../../../../../utils/callServiceMethod';
-import { getSessionIdFromCookies } from '../../../../../../utils/session';
 import OptionsQuestionSummary from '../../components/OptionsQuestionSummary';
-import {
-  getErrorPageParams,
-  questionErrorPageRedirect,
-} from './editQuestionServiceError';
+import InferProps from '../../../../../../types/InferProps';
+import getServerSideProps from './question-options.getServerSideProps';
 
-export const getServerSideProps: GetServerSideProps = async ({
-  params,
-  req,
-  res,
-}) => {
-  const sessionId = getSessionIdFromCookies(req);
-  const { applicationId, sectionId, questionId } = params as Record<
-    string,
-    string
-  >;
-
-  let fieldErrors: ValidationError[] = [];
-  let options: string[] = [''];
-  let applicationFormSummary: ApplicationFormSummary;
-  let questionData: ApplicationFormQuestion;
-  let questionSummary: QuestionWithOptionsSummary;
-
-  try {
-    applicationFormSummary = await getApplicationFormSummary(
-      applicationId,
-      getSessionIdFromCookies(req)
-    );
-
-    if (applicationFormSummary.applicationStatus === 'PUBLISHED') {
-      return {
-        redirect: {
-          destination: `/build-application/${applicationId}/${sectionId}/${questionId}/preview`,
-          statusCode: 302,
-        },
-      };
-    }
-    questionData = await getQuestion(
-      sessionId,
-      applicationId,
-      sectionId,
-      questionId
-    );
-
-    questionSummary = {
-      fieldTitle: questionData.fieldTitle,
-      hintText: questionData.hintText,
-      optional: (!questionData.validation.mandatory).toString(),
-      responseType: questionData.responseType,
-    };
-  } catch (err) {
-    return questionErrorPageRedirect(applicationId);
-  }
-
-  const sectionName = applicationFormSummary.sections.find(
-    (section) => section.sectionId === sectionId
-  )?.sectionTitle;
-
-  const result = await callServiceMethod(
-    req,
-    res,
-    async (body: any) => {
-      options = Object.keys(body).reduce((array, key) => {
-        if (key.startsWith('options')) {
-          array.push(...body[key]);
-        }
-
-        if (key.startsWith('delete_')) {
-          const deleteOptionIndex = Number(key.split('_')[1]);
-          array.splice(deleteOptionIndex, 1);
-        }
-        return array;
-      }, [] as string[]);
-
-      switch (true) {
-        case 'add-another-option' in body:
-          options.push('');
-
-          return {
-            data: 'OPTION_ADDED',
-          };
-
-        case 'save-and-continue' in body:
-          await patchQuestion(sessionId, applicationId, sectionId, questionId, {
-            options: options,
-          });
-
-          return {
-            data: 'QUESTION_SAVED',
-          };
-
-        default:
-          return {
-            data: 'OPTION_REMOVED',
-          };
-      }
-    },
-    (response: { data: string }) => {
-      return response.data === 'QUESTION_SAVED'
-        ? `/build-application/${applicationId}/dashboard`
-        : '';
-    },
-    getErrorPageParams(applicationId)
-  );
-
-  if ('redirect' in result) {
-    if (result.redirect.destination !== '') {
-      return result;
-    }
-  } else if ('body' in result) {
-    //The backend currently returns 2 forms of errors, Class level or field level
-    fieldErrors = result.fieldErrors
-      .reduce((array, error) => {
-        //To remain GDS compliant we need to link each error to the fields which caused the validation error.
-        //here we catch class level errors and return individual errors to link to each field
-        if (error.fieldName === 'options') {
-          const newElementErrors = options.map(
-            (_option, index) =>
-              ({
-                fieldName: `options[${index}]`,
-                errorMessage: error.errorMessage,
-              } as ValidationError)
-          );
-          array.push(...newElementErrors);
-        } else if (error.fieldName.startsWith('options')) {
-          //Here we catch field level errors and update the backend fieldName to match the frontend fieldName
-          array.push({
-            fieldName: error.fieldName.split('.')[0],
-            errorMessage: error.errorMessage,
-          });
-        }
-        return array;
-      }, [] as ValidationError[])
-      .sort((a, b) =>
-        a.fieldName.localeCompare(b.fieldName, undefined, { numeric: true })
-      );
-  }
-  if (req.method !== 'POST') {
-    options = questionData.options as string[];
-  }
-  const { publicRuntimeConfig } = getConfig();
-  return {
-    props: {
-      pageCaption: sectionName,
-      questionSummary,
-      backButtonHref: `/build-application/${applicationId}/${sectionId}/${questionId}/edit/question-content`,
-      formAction: `${publicRuntimeConfig.SUB_PATH}/build-application/${applicationId}/${sectionId}/${questionId}/edit/question-options`,
-      fieldErrors,
-      cancelChangesHref: `/build-application/${applicationId}/${sectionId}/${questionId}/preview`,
-      options,
-      csrfToken: res.getHeader('x-csrf-token') as string,
-    },
-  };
-};
+export { getServerSideProps };
 
 const QuestionOptions = ({
   pageCaption,
@@ -185,7 +17,7 @@ const QuestionOptions = ({
   fieldErrors,
   options,
   csrfToken,
-}: QuestionOptionProps) => {
+}: InferProps<typeof getServerSideProps>) => {
   return (
     <>
       <Meta
@@ -264,17 +96,6 @@ const QuestionOptions = ({
       </div>
     </>
   );
-};
-
-type QuestionOptionProps = {
-  pageCaption: string;
-  questionSummary: QuestionWithOptionsSummary;
-  backButtonHref: string;
-  formAction: string;
-  cancelChangesHref: string;
-  fieldErrors: ValidationError[];
-  options: string[];
-  csrfToken: string;
 };
 
 export default QuestionOptions;
